+++ ucommerce_client/lib/main.dart
@@ -9,6 +9,6 @@ class MyApp extends StatelessWidget {
   const MyApp({Key? key}) : super(key: key);
   @override
   Widget build(BuildContext context) {
-    return const MaterialApp(debugShowCheckedModeBanner: false, home: Admintiendas());
+    return const MaterialApp(debugShowCheckedModeBanner: false, home: Compra());
   }
 }
